Remove UghHelper class
