--- Kitodo/src/main/java/org/kitodo/production/helper/UghHelper.java
@@ -1,91 +1,0 @@
-/*
- * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
- *
- * This file is part of the Kitodo project.
- *
- * It is licensed under GNU General Public License version 3 or later.
- *
- * For the full copyright and license information, please read the
- * GPL3-License.txt file that was distributed with this source code.
- */
-
-package org.kitodo.production.helper;
-
-import java.util.List;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.kitodo.data.database.beans.Process;
-import org.kitodo.exceptions.UghHelperException;
-import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyDocStructHelperInterface;
-import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyMetadataHelper;
-import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyMetadataTypeHelper;
-import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyPrefsHelper;
-import org.kitodo.production.services.ServiceManager;
-
-public class UghHelper {
-    private static final Logger logger = LogManager.getLogger(UghHelper.class);
-
-    /**
-     * Private constructor to hide the implicit public one.
-     */
-    private UghHelper() {
-
-    }
-
-    /**
-     * MetadataType aus Preferences eines Prozesses ermitteln.
-     *
-     * @param inProzess
-     *            Process object
-     * @param inName
-     *            String
-     * @return MetadataType
-     */
-    public static LegacyMetadataTypeHelper getMetadataType(Process inProzess, String inName) throws UghHelperException {
-        LegacyPrefsHelper myPrefs = ServiceManager.getRulesetService().getPreferences(inProzess.getRuleset());
-        return getMetadataType(myPrefs, inName);
-    }
-
-    /**
-     * MetadataType aus Preferences ermitteln.
-     *
-     * @param inPrefs
-     *            Prefs object
-     * @param inName
-     *            String
-     * @return MetadataType
-     */
-    public static LegacyMetadataTypeHelper getMetadataType(LegacyPrefsHelper inPrefs, String inName) throws UghHelperException {
-        LegacyMetadataTypeHelper mdt = inPrefs.getMetadataTypeByName(inName);
-        if (mdt == null) {
-            throw new UghHelperException("MetadataType does not exist in current Preferences: " + inName);
-        }
-        return mdt;
-    }
-
-    /**
-     * Metadata eines Docstructs ermitteln.
-     *
-     * @param inStruct
-     *            DocStruct object
-     * @param inMetadataType
-     *            MetadataType object
-     * @return Metadata
-     */
-    public static LegacyMetadataHelper getMetadata(LegacyDocStructHelperInterface inStruct, LegacyMetadataTypeHelper inMetadataType) {
-        if (inStruct != null && inMetadataType != null) {
-            List<? extends LegacyMetadataHelper> all = inStruct.getAllMetadataByType(inMetadataType);
-            if (all.isEmpty()) {
-                LegacyMetadataHelper md = new LegacyMetadataHelper(inMetadataType);
-                md.setDocStruct(inStruct);
-                inStruct.addMetadata(md);
-                return md;
-            } else {
-                return all.get(0);
-            }
-        }
-        return null;
-    }
-
-}
