+++ Kitodo/src/main/java/org/kitodo/production/helper/metadata/legacytypeimplementations/LegacyPrefsHelper.java
@@ -27,6 +27,7 @@ import org.kitodo.api.dataeditor.rulesetmanagement.MetadataViewWithValuesInterfa
 import org.kitodo.api.dataeditor.rulesetmanagement.RulesetManagementInterface;
 import org.kitodo.api.dataeditor.rulesetmanagement.StructuralElementViewInterface;
 import org.kitodo.data.database.beans.User;
+import org.kitodo.exceptions.UghHelperException;
 import org.kitodo.production.helper.Helper;
 import org.kitodo.production.metadata.MetadataProcessor;
 import org.kitodo.production.services.ServiceManager;
@@ -76,6 +77,24 @@ public class LegacyPrefsHelper {
                     priorityList);
                 return new LegacyLogicalDocStructTypeHelper(divisionView);
         }
+    }
+
+    /**
+     * MetadataType aus Preferences ermitteln.
+     *
+     * @param inPrefs
+     *            Prefs object
+     * @param inName
+     *            String
+     * @return MetadataType
+     */
+    @Deprecated
+    public static LegacyMetadataTypeHelper getMetadataType(LegacyPrefsHelper inPrefs, String inName) throws UghHelperException {
+        LegacyMetadataTypeHelper mdt = inPrefs.getMetadataTypeByName(inName);
+        if (mdt == null) {
+            throw new UghHelperException("MetadataType does not exist in current Preferences: " + inName);
+        }
+        return mdt;
     }
 
     /**
